mutil datasource config

+++ cloud-routing-datasource/src/main/java/com/barry/cloud/platform/routing/entity/Staff.java
@@ -9,7 +9,7 @@ public class Staff {
 
     private String name;
 
-    private Byte gender;
+    private Short gender;
 
     private String mobile;
